rename to app.py, display images in rows of 2
